# lib/api-zod/src/generated/types/heritageSite.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeritageSiteSourceType } from './heritageSiteSourceType';

export interface HeritageSite {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  shortDescription: string | null;
  country: string;
  /** @nullable */
  state: string | null;
  /** @nullable */
  region: string | null;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  category: string | null;
  /** @nullable */
  yearInscribed: number | null;
  /** @nullable */
  criteria: string | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  imageAuthor: string | null;
  /** @nullable */
  imageCopyright: string | null;
  /** @nullable */
  imageCaption: string | null;
  /** @nullable */
  components: string | null;
  sourceName: string;
  sourceUrl: string;
  sourceType: HeritageSiteSourceType;
}